+++ lib/vagrant/util/downloader.rb
@@ -138,11 +138,14 @@ module Vagrant
           # If we already retried, raise it.
           raise if retried
 
+          @logger.error("Exit code: #{e.extra_data[:code]}")
+
           # If its any error other than 33, it is an error.
-          raise if e.extra_data[:exit_code].to_i != 33
+          raise if e.extra_data[:code].to_i != 33
 
           # Exit code 33 means that the server doesn't support ranges.
           # In this case, try again without resume.
+          @logger.error("Error is server doesn't support byte ranges. Retrying from scratch.")
           @continue = false
           retried = true
           retry